fixed some of static analysis warnings reported here: https://horde.devtools.epicgames.com/job/6398595e2960b73220e958d3?step=1b3d #preflight 6398680d35203bc7aa5bd7ec
[CL 23490169 by Andriy Tylychko in ue5-main branch]

+++ Engine/Source/Runtime/Core/Public/Windows/WindowsSemaphore.h
@@ -33,7 +33,7 @@ public:
 	bool TryAcquire(FTimespan Timeout = FTimespan::Zero())
 	{
 		HRESULT hRes = WaitForSingleObject(Semaphore, (DWORD)Timeout.GetTotalMilliseconds());
-		checkfSlow(hRes == WAIT_OBJECT_0 || hRes == WAIT_TIMEOUT, TEXT("Acquiring semaphore failed: %d (%d)"), hRes, GetLastError());
+		checkfSlow(hRes == WAIT_OBJECT_0 || hRes == WAIT_TIMEOUT, TEXT("Acquiring semaphore failed: %d (%d)"), (int)hRes, GetLastError());
 		return hRes == WAIT_OBJECT_0;
 	}
